Introduces NPM and Webpack for JS builds, and adds a bit of organization.
This feature isn't done. Right now I have things set up to commit the compiled
JS bundles. I'm still figuring out how I want to organize this project, so I'll
come back to it.
