+++ jekyll-site-src/README.md
@@ -1,0 +1,23 @@
+# jekyll-site-src
+
+### Prerequisites for development
+
+- node v7+
+
+### Installation
+
+```sh
+$ npm i
+```
+
+### Building JavaScript for Development
+
+The following will start a watch that builds the JavaScript whenever it changes:
+```sh
+$ npm run dev
+```
+
+### Building JavaScript for Production
+```sh
+$ npm run build
+```
