+++ jekyll-site-src/_config.yml
@@ -31,7 +31,11 @@ kramdown:
 #    css: style
 
 # Build target
-exclude: [build.sh]
+exclude:
+  - build.sh
+  - node_modules
+  - package.json
+  - webpack.config.js
 
 defaults:
 #layout specification
