+++ jekyll-site-src/_includes/head.html
@@ -13,25 +13,9 @@
   <meta property="og:site_name" content="{{ site.title | escape }}" />
   <meta property="article:section" content="{% if page.title %}{{ page.title | escape }}{% endif %}" />
 
-  <link href='//fonts.googleapis.com/css?family=Roboto:400,300,500|Material+Icons' rel='stylesheet' type='text/css'>
+  <link href='//fonts.googleapis.com/css?family=Roboto:300,400,500|Material+Icons' rel='stylesheet' type='text/css'>
   <link rel="stylesheet" href="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/lib/codemirror.css">
   <link rel="stylesheet" href="{{ site.url }}{{ site.folder }}/css/main.css">
   <link rel="canonical" href="{{ page.url | replace:'index.html','' | prepend: site.folder | prepend: site.url }}">
   <link rel='shortcut icon' type='image/x-icon' href='{{ site.url }}{{ site.folder }}/favicon.ico' />
-  <script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/lib/codemirror.js"></script>
-  <script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/clike/clike.js"></script>
-  <script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/swift/swift.js"></script>
-  <script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/shell/shell.js"></script>
-  <script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/ruby/ruby.js"></script>
-  <script src='{{ site.url }}{{ site.folder }}/js/mdccoderender.js'></script>
-
-  <script>
-    (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-    (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-    })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-
-    ga('create', '{{ site.analytics_id }}', 'auto');
-    ga('send', 'pageview');
-  </script>
 </head>
