+++ jekyll-site-src/_includes/scripts.html
@@ -1,0 +1,16 @@
+<script src='{{ site.url }}{{ site.folder }}/js/index.bundle.js'></script>
+<script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/lib/codemirror.js"></script>
+<script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/clike/clike.js"></script>
+<script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/swift/swift.js"></script>
+<script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/shell/shell.js"></script>
+<script src="{{ site.url }}{{ site.folder }}/thirdparty/codemirror/mode/ruby/ruby.js"></script>
+
+<script>
+  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
+  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+  })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
+
+  ga('create', '{{ site.analytics_id }}', 'auto');
+  ga('send', 'pageview');
+</script>
