+++ jekyll-site-src/_js_src/index.js
@@ -1,0 +1,1 @@
+import './mdc-code-render';
