+++ jekyll-site-src/_layouts/default.html
@@ -1,10 +1,7 @@
 <!DOCTYPE html>
 <html>
-
   {% include head.html %}
-
   <body>
-
     {% include header.html %}
     {% include sidebar.html %}
 
@@ -13,7 +10,7 @@
     </section>
 
     {% include footer.html %}
-
     {% include frontend_templates.html %}
+    {% include scripts.html %}
   </body>
 </html>
