+++ jekyll-site-src/_layouts/detail.html
@@ -1,10 +1,7 @@
 <!DOCTYPE html>
 <html>
-
   {% include head.html %}
-
   <body class="{{ page.layout }} {{ page.section }}">
-
     {% include header.html %}
     {% include sidebar.html %}
 
@@ -15,7 +12,7 @@
     </div>
 
     {% include footer.html %}
-
     {% include frontend_templates.html %}
+    {% include scripts.html %}
   </body>
 </html>
