+++ jekyll-site-src/_layouts/homepage.html
@@ -1,10 +1,7 @@
 <!DOCTYPE html>
 <html>
-
   {% include head.html %}
-
   <body id="homepage">
-
     {% include header.html %}
 
     <div id="main_content">
@@ -14,7 +11,7 @@
     </div>
 
     {% include footer.html %}
-
     {% include frontend_templates.html %}
+    {% include scripts.html %}
   </body>
 </html>
